requirements.txt
streamlit
pandas
numpy
python-dateutil
pdfplumber
matplotlib
